requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
chromadb>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
